move controllers into subfolder (in preparation of custom hooks)
