+++ src/components/Viewer/Viewer.tsx
@@ -2,8 +2,8 @@ import { ResizeEvent, ViewerEvent } from "openseadragon";
 import { useCallback, useEffect, useRef } from "react";
 
 import { useBoundStore } from "../../store/boundStore";
-import OpenSeadragonController from "./OpenSeadragonController";
-import WebGLController from "./WebGLController";
+import OpenSeadragonController from "./controllers/OpenSeadragonController";
+import WebGLController from "./controllers/WebGLController";
 
 export default function Viewer() {
   const osRef = useRef<OpenSeadragonController | null>(null);
